create port and search service
