+++ src/main/resources/config/application.yml
@@ -150,10 +150,10 @@ base-url:
 #================================================================
 # Keycloak Server Configurations
 keycloak:
-  realm: demo_realm
+  realm: ravi_realm
   credentials: 
-    secret: zf8GS1Lc2lCzIUULNIwyateNNFJrCxcj
-  resource: Ingress_Microservice
+    secret: 2MGBo1s8upyPn8lRsHW0scW6WkrLOj8K
+  resource: search-as-a-service-egress
   auth-server-url: http://localhost:8080/auth
   ssl-required: external
   public-client: false
